Add a Role to handle paths for path find functionality
This Role adds an attribute "path" to a Result, giving it the ability to
build a directory path based on the data for a given lane. It might well
not need to be a Role, since it is probably only applicable to a Lane,
but I haven't thought through all of the use-cases yet.