Inline code from HasPath into LatestLane
At first I thought that the code to build a path might be more widely
applicable, but it looks as though it's only needed by the LatestLane
Result. Move the code into the Result class and delete the Role.

Also, this commit fixes a bug (cargo-culted from the original VRTrack
class on which the HasPath code was patterned) in the code that gets the
genus and species for a sample. In the original code it was assumed that
there would also be a genus and species, separated by a space. Turns out
no. The regex that splits the genus-and-species string now takes into
account the possibility that there's only a genus.