Add a top-level schema that loads selected results
Loading all of the Result classes slows things down a lot. This commit
adds a new ReducedSchema class, which used load_classes() instead of
load_namespaces(), allowing us to specify which Result classes get
loaded.

Since load_classes doesn't automatically register ResultSets with their
corresponding Results, we need to add a call to Result::LatestLane to do
that.

Since the only method on Schema is now duplicated, moved it to a Role
and applied that to both schema classes.