De-moosify and restrict class loading
Convert all but one Result class into regular perl classes, as opposed
to moose classes. LatestLane is still a moose class because it uses some
moose functionality.

Also, restrict the Schema class to loading only those classes that
are required by Bio-Path-Find. It will probably be necessary to expand
the list of loaded classes, or to add a "FullSchema" class which loads
everything.